+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This is a *milestone* (i.e. still in development) version made available to try 
   * Updated to Groovy 3 - scripts now support more recent Java syntax (e.g. lambdas, try-with-resources)
   * Pasting files results in them being converted to absolute paths
   * New 'Paste & escape' command to automatically escape characters for Java Strings
+* Scriptable 'Select objects by classification' command
 * Enable files to be pasted into the script editor (converting them to absolute paths)
 * Updated methods to save/load points within the counting tool
   * Use TSV files to improve portability